@@ -1,13 +1,13 @@
 plugins {
     id 'java'
-    id 'org.springframework.boot' version '2.7.2'
+    id 'org.springframework.boot' version '3.0.5'
     id 'io.spring.dependency-management' version '1.0.12.RELEASE'
     id "org.asciidoctor.jvm.convert" version "3.3.2"
 }
 
 group = 'com.kakjziblog'
 version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '11'
+sourceCompatibility = '17'
 
 configurations {
     compileOnly {
@@ -20,7 +20,6 @@ repositories {
     mavenCentral()
 }
 ext {
-    asciidocVersion = "2.0.6.RELEASE"
     snippetsDir = file('build/generated-snippets')
 }
 
@@ -32,15 +31,14 @@ dependencies {
     implementation 'org.springframework.security:spring-security-crypto'
     implementation 'org.bouncycastle:bcprov-jdk15on:1.70'
 
-    implementation 'com.querydsl:querydsl-core'
-    implementation 'com.querydsl:querydsl-jpa'
-
-    asciidoctorExt "org.springframework.restdocs:spring-restdocs-asciidoctor:${asciidocVersion}"
-    testImplementation "org.springframework.restdocs:spring-restdocs-mockmvc:${asciidocVersion}"
-
-    annotationProcessor "com.querydsl:querydsl-apt:5.0.0:jpa"
+    implementation 'com.querydsl:querydsl-core:5.0.0'
+    implementation 'com.querydsl:querydsl-jpa:5.0.0:jakarta'
+    annotationProcessor "com.querydsl:querydsl-apt:5.0.0:jakarta"
     annotationProcessor 'jakarta.persistence:jakarta.persistence-api'
     annotationProcessor 'jakarta.annotation:jakarta.annotation-api'
+
+    asciidoctorExt "org.springframework.restdocs:spring-restdocs-asciidoctor:3.0.0"
+    testImplementation "org.springframework.restdocs:spring-restdocs-mockmvc:3.0.0"
 
     compileOnly 'org.projectlombok:lombok'
     developmentOnly 'org.springframework.boot:spring-boot-devtools'
@@ -50,9 +48,7 @@ dependencies {
 
     implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
     runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.11.5'
-            // Uncomment the next line if you want to use RSASSA-PSS (PS256, PS384, PS512) algorithms:
-            //'org.bouncycastle:bcprov-jdk15on:1.70',
-    runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.11.5' // or 'io.jsonwebtoken:jjwt-gson:0.11.5' for gson
+    runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.11.5'
 }
 tasks.named('test') {
     useJUnitPlatform()
