@@ -1,6 +1,6 @@
 package com.kakjziblog.api.request;
 
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotBlank;
 
 import lombok.Builder;
 import lombok.Getter;
