@@ -1,7 +1,7 @@
 package com.kakjziblog.api.request;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 
 import com.kakjziblog.api.domain.Category;
 
